generate questions updated
